Improved structure of panels-vs-controls
