+++ assets/src/edit-story/components/designPanel/controls/size/useUpdateWidth.js
